refactor: moved scripts into tools/
